+++ lib/data/mapper/feels_like_mapper.dart
@@ -2,9 +2,9 @@ import 'package:hello_weather/data/api/model/feels_like_api.dart';
 
 class FeelsLikeMapper {
   static Map<String, double> fromApi(ApiFeelsLike list) => {
-        "day": list.day.toDouble(),
-        "night": list.night.toDouble(),
-        "evening": list.evening.toDouble(),
-        "morning": list.morning.toDouble(),
+        "day": list.day,
+        "night": list.night,
+        "evening": list.evening,
+        "morning": list.morning,
       };
 }
